CEA parsing no longer header only
