+++ src/Makefile.am
@@ -18,6 +18,8 @@ libantioch_la_SOURCES += core/src/chemical_mixture.C
 # parsing
 libantioch_la_SOURCES += parsing/src/ascii_parser.C
 libantioch_la_SOURCES += parsing/src/blottner_parsing.C
+libantioch_la_SOURCES += parsing/src/cea_mixture_parsing.C
+libantioch_la_SOURCES += parsing/src/cea_mixture_ascii_parsing.C
 libantioch_la_SOURCES += parsing/src/transport_species_parsing.C
 
 # transport
